requirements.txt
jittor==1.3.1.18
matplotlib==3.10.3
numpy==2.3.0
Pillow==11.2.1
scipy==1.15.3
skimage==0.0
tqdm==4.67.1
